Final commit CSS and HTML Update
Added the the border for the table. Made the table header black, the content blue and added font weight normal. Made the table data vertically align with the top of the table. Added column widths. Added the thead height. Added font weight and made background transparent of the heading of the table. Added the rest of the comment for the code. Ready for review.